+++ CHANGELOG.md
@@ -5,6 +5,8 @@ v0.1.9 (unreleased)
 
 - Tune producer max message bytes to 10mb.
 - Tune consumer default fetch bytes to 10mb.
+- Remove Offset.Initial.Reset config since it is unused.
+- Add Offset.Commit.Enabled config to enable auto offset commit.
 
 
 v0.1.8 (2018-06-04)